dasClangBind: regen bindings for libclang 22.1.5, enable on mingw CI
Bumps the bindings from their libclang-16 baseline to libclang-22.1.5 -
matches what apt.llvm.org's llvm-toolchain-noble-22 ships for the
linux extended_checks self-binder step, and the same LLVM 22.1.5
aleksisch/llvm-shared-builds tarball dasLLVM already uses.

Changes:
- Regenerated all of modules/dasClangBind/src/*.inc + func_*.cpp
  (now 18 chunks vs 17) against LLVM 22.1.5 headers. Picks up the
  CXCursor_OMPArraySectionExpr -> CXCursor_ArraySectionExpr rename
  and ~340 other API additions/changes between 16 and 22.
- find_package(Clang 16.0.6) -> find_package(Clang 22.1) in
  modules/dasClangBind/CMakeLists.txt. Matches both msys2 clang64
  (22.1.4) and apt.llvm.org's noble-22 package via the standard
  ConfigVersion compatibility check.
- POST_BUILD copy of libclang.dll + transitive runtime deps next to
  daslang.exe (mirrors the dasHV/dasGlfw pattern from #2838/#2840).
  Required because daslang loads .shared_module files via
  LoadLibraryEx with restricted search flags (PATH is NOT searched -
  see src/misc/sysos.cpp:loadDynamicLibrary). Multi-config /
  single-config split for symmetry. MSVC path uses official LLVM's
  self-contained libclang.dll; mingw needs the full chain
  (libclang + libLLVM-N + libffi/zlib/zstd/libxml2/libiconv + libc++
  + libunwind).
- Broadened the ToBasicType<unsigned long> specialization in
  include/daScript/ast/ast_typedecl.h from #if defined(_MSC_VER)
  to #if defined(_WIN32). Without it, mingw fails to build the
  binding for CXUnsavedFile::Length - Windows LLP64 makes
  'unsigned long' a distinct 32-bit type regardless of compiler.
- .github/workflows/extended_checks.yml: dropped libclang-16-dev,
  llvm-16-dev, clang-16 from the apt install list (no longer
  needed); switched the bind_clangbind self-binder --clang_path from
  /usr/lib/llvm-16/include/ to /usr/lib/llvm-22/include/.
- .github/workflows/build.yml: dropped -DDAS_CLANG_BIND_DISABLED=ON
  from the build_windows_mingw cmake invocation. find_package(Clang
  22.1) now matches the msys2 clang64 sysroot so the module builds
  cleanly under mingw.

Locally verified: daslang modules/dasClangBind/examples/browse.das
walks sample.h and generates daslang bindings end-to-end on mingw.
CI's linux extended_checks self-binder step is the byte-canonical
re-check.
