+++ .github/workflows/build.yml
@@ -482,9 +482,9 @@ jobs:
   # (windows-latest-fat), separate top-level job so mingw churn doesn't risk
   # the much-larger MSVC entries.
   #
-  # dasClangBind stays disabled here — pinned to libclang 16.0.6 (find_package
-  # strict version), which the msys2 clang64 sysroot doesn't ship. Tracked
-  # separately as the bind_clangbind regen followup.
+  # dasClangBind is enabled here — find_package(Clang 22.1) matches msys2's
+  # libclang 22.1.x and POST_BUILD copies the libclang+libLLVM+libc++ chain
+  # next to daslang.exe so the .shared_module dlopen resolves at runtime.
   ###########################################################
     needs: pre_job
     if: needs.pre_job.outputs.should_skip != 'true'
@@ -518,7 +518,7 @@ jobs:
           -DCMAKE_BUILD_TYPE=Release \
           -DCMAKE_C_COMPILER=clang \
           -DCMAKE_CXX_COMPILER=clang++ \
-          -DDAS_CLANG_BIND_DISABLED=ON \
+          -DDAS_CLANG_BIND_DISABLED=OFF \
           -DDAS_LLVM_DISABLED=OFF
         cmake --build ./build-mingw --parallel
