+++ modules/dasClangBind/src/dasClangBind.alias.add.inc
@@ -16,93 +16,103 @@ auto alias_CXDiagnosticSet = typeFactory<void *>::make(lib);
 alias_CXDiagnosticSet->alias = "CXDiagnosticSet";
 addAlias(alias_CXDiagnosticSet);
 
-// from clang-c/Index.h:76:15
+// from clang-c/Index.h:80:15
 auto alias_CXIndex = typeFactory<void *>::make(lib);
 alias_CXIndex->alias = "CXIndex";
 addAlias(alias_CXIndex);
 
-// from clang-c/Index.h:82:34
+// from clang-c/Index.h:86:34
 auto alias_CXTargetInfo = typeFactory<CXTargetInfoImpl *>::make(lib);
 alias_CXTargetInfo->alias = "CXTargetInfo";
 addAlias(alias_CXTargetInfo);
 
-// from clang-c/Index.h:87:39
+// from clang-c/Index.h:91:39
 auto alias_CXTranslationUnit = typeFactory<CXTranslationUnitImpl *>::make(lib);
 alias_CXTranslationUnit->alias = "CXTranslationUnit";
 addAlias(alias_CXTranslationUnit);
 
-// from clang-c/Index.h:93:15
+// from clang-c/Index.h:97:15
 auto alias_CXClientData = typeFactory<void *>::make(lib);
 alias_CXClientData->alias = "CXClientData";
 addAlias(alias_CXClientData);
 
-// from clang-c/Index.h:2423:33
+// from clang-c/Index.h:2657:33
 auto alias_CXCursorSet = typeFactory<CXCursorSetImpl *>::make(lib);
 alias_CXCursorSet->alias = "CXCursorSet";
 addAlias(alias_CXCursorSet);
 
-// from clang-c/Index.h:3671:35
+// from clang-c/Index.h:4006:35
 auto alias_CXCursorVisitor = typeFactory<CXChildVisitResult (*)(CXCursor, CXCursor, void *)>::make(lib);
 alias_CXCursorVisitor->alias = "CXCursorVisitor";
 addAlias(alias_CXCursorVisitor);
 
-// from clang-c/Index.h:3812:15
+// from clang-c/Index.h:4049:37
+auto alias_CXCursorVisitorBlock = typeFactory<_CXChildVisitResult *>::make(lib);
+alias_CXCursorVisitorBlock->alias = "CXCursorVisitorBlock";
+addAlias(alias_CXCursorVisitorBlock);
+
+// from clang-c/Index.h:4147:15
 auto alias_CXPrintingPolicy = typeFactory<void *>::make(lib);
 alias_CXPrintingPolicy->alias = "CXPrintingPolicy";
 addAlias(alias_CXPrintingPolicy);
 
-// from clang-c/Index.h:4159:15
+// from clang-c/Index.h:4637:15
 auto alias_CXModule = typeFactory<void *>::make(lib);
 alias_CXModule->alias = "CXModule";
 addAlias(alias_CXModule);
 
-// from clang-c/Index.h:4659:15
+// from clang-c/Index.h:5182:15
 auto alias_CXCompletionString = typeFactory<void *>::make(lib);
 alias_CXCompletionString->alias = "CXCompletionString";
 addAlias(alias_CXCompletionString);
 
-// from clang-c/Index.h:5464:16
+// from clang-c/Index.h:5987:16
 auto alias_CXInclusionVisitor = typeFactory<void (*)(void *, CXSourceLocation *, unsigned int, void *)>::make(lib);
 alias_CXInclusionVisitor->alias = "CXInclusionVisitor";
 addAlias(alias_CXInclusionVisitor);
 
-// from clang-c/Index.h:5494:15
+// from clang-c/Index.h:6017:15
 auto alias_CXEvalResult = typeFactory<void *>::make(lib);
 alias_CXEvalResult->alias = "CXEvalResult";
 addAlias(alias_CXEvalResult);
 
-// from clang-c/Index.h:5565:15
-auto alias_CXRemapping = typeFactory<void *>::make(lib);
-alias_CXRemapping->alias = "CXRemapping";
-addAlias(alias_CXRemapping);
+// from clang-c/Index.h:6145:47
+auto alias_CXCursorAndRangeVisitorBlock = typeFactory<_CXCursorAndRangeVisitorBlock *>::make(lib);
+alias_CXCursorAndRangeVisitorBlock->alias = "CXCursorAndRangeVisitorBlock";
+addAlias(alias_CXCursorAndRangeVisitorBlock);
 
-// from clang-c/Index.h:5698:15
+// from clang-c/Index.h:6159:15
 auto alias_CXIdxClientFile = typeFactory<void *>::make(lib);
 alias_CXIdxClientFile->alias = "CXIdxClientFile";
 addAlias(alias_CXIdxClientFile);
 
-// from clang-c/Index.h:5703:15
+// from clang-c/Index.h:6164:15
 auto alias_CXIdxClientEntity = typeFactory<void *>::make(lib);
 alias_CXIdxClientEntity->alias = "CXIdxClientEntity";
 addAlias(alias_CXIdxClientEntity);
 
-// from clang-c/Index.h:5709:15
+// from clang-c/Index.h:6170:15
 auto alias_CXIdxClientContainer = typeFactory<void *>::make(lib);
 alias_CXIdxClientContainer->alias = "CXIdxClientContainer";
 addAlias(alias_CXIdxClientContainer);
 
-// from clang-c/Index.h:5715:15
+// from clang-c/Index.h:6176:15
 auto alias_CXIdxClientASTFile = typeFactory<void *>::make(lib);
 alias_CXIdxClientASTFile->alias = "CXIdxClientASTFile";
 addAlias(alias_CXIdxClientASTFile);
 
-// from clang-c/Index.h:6130:15
+// from clang-c/Index.h:6591:15
 auto alias_CXIndexAction = typeFactory<void *>::make(lib);
 alias_CXIndexAction->alias = "CXIndexAction";
 addAlias(alias_CXIndexAction);
 
-// from clang-c/Index.h:6282:32
+// from clang-c/Index.h:6743:32
 auto alias_CXFieldVisitor = typeFactory<CXVisitorResult (*)(CXCursor, void *)>::make(lib);
 alias_CXFieldVisitor->alias = "CXFieldVisitor";
 addAlias(alias_CXFieldVisitor);
+
+// from clang-c/Index.h:6961:15
+auto alias_CXRemapping = typeFactory<void *>::make(lib);
+alias_CXRemapping->alias = "CXRemapping";
+addAlias(alias_CXRemapping);
