+++ modules/dasClangBind/src/dasClangBind.enum.add.inc
@@ -7,6 +7,7 @@ addEnumeration(new Enumeration_CXLoadDiag_Error());
 addEnumeration(new Enumeration_CXDiagnosticDisplayOptions());
 addEnumeration(new Enumeration_CXAvailabilityKind());
 addEnumeration(new Enumeration_CXCursor_ExceptionSpecificationKind());
+addEnumeration(new Enumeration_CXChoice());
 addEnumeration(new Enumeration_CXGlobalOptFlags());
 addEnumeration(new Enumeration_CXTranslationUnit_Flags());
 addEnumeration(new Enumeration_CXSaveTranslationUnit_Flags());
@@ -26,6 +27,7 @@ addEnumeration(new Enumeration_CXTypeLayoutError());
 addEnumeration(new Enumeration_CXRefQualifierKind());
 addEnumeration(new Enumeration_CX_CXXAccessSpecifier());
 addEnumeration(new Enumeration_CX_StorageClass());
+addEnumeration(new Enumeration_CX_BinaryOperatorKind());
 addEnumeration(new Enumeration_CXChildVisitResult());
 addEnumeration(new Enumeration_CXPrintingPolicyProperty());
 addEnumeration(new Enumeration_CXObjCPropertyAttrKind());
@@ -47,3 +49,5 @@ addEnumeration(new Enumeration_CXIdxObjCContainerKind());
 addEnumeration(new Enumeration_CXIdxEntityRefKind());
 addEnumeration(new Enumeration_CXSymbolRole());
 addEnumeration(new Enumeration_CXIndexOptFlags());
+addEnumeration(new Enumeration_CXBinaryOperatorKind());
+addEnumeration(new Enumeration_CXUnaryOperatorKind());
