+++ modules/dasClangBind/src/dasClangBind.enum.decl.cast.inc
@@ -8,6 +8,7 @@ DAS_BIND_ENUM_CAST(CXLoadDiag_Error);
 DAS_BIND_ENUM_CAST(CXDiagnosticDisplayOptions);
 DAS_BIND_ENUM_CAST(CXAvailabilityKind);
 DAS_BIND_ENUM_CAST(CXCursor_ExceptionSpecificationKind);
+DAS_BIND_ENUM_CAST(CXChoice);
 DAS_BIND_ENUM_CAST(CXGlobalOptFlags);
 DAS_BIND_ENUM_CAST(CXTranslationUnit_Flags);
 DAS_BIND_ENUM_CAST(CXSaveTranslationUnit_Flags);
@@ -27,6 +28,7 @@ DAS_BIND_ENUM_CAST(CXTypeLayoutError);
 DAS_BIND_ENUM_CAST(CXRefQualifierKind);
 DAS_BIND_ENUM_CAST(CX_CXXAccessSpecifier);
 DAS_BIND_ENUM_CAST(CX_StorageClass);
+DAS_BIND_ENUM_CAST(CX_BinaryOperatorKind);
 DAS_BIND_ENUM_CAST(CXChildVisitResult);
 DAS_BIND_ENUM_CAST(CXPrintingPolicyProperty);
 DAS_BIND_ENUM_CAST(CXObjCPropertyAttrKind);
@@ -48,3 +50,5 @@ DAS_BIND_ENUM_CAST(CXIdxObjCContainerKind);
 DAS_BIND_ENUM_CAST(CXIdxEntityRefKind);
 DAS_BIND_ENUM_CAST(CXSymbolRole);
 DAS_BIND_ENUM_CAST(CXIndexOptFlags);
+DAS_BIND_ENUM_CAST(CXBinaryOperatorKind);
+DAS_BIND_ENUM_CAST(CXUnaryOperatorKind);
