+++ modules/dasClangBind/src/dasClangBind.func.decl.inc
@@ -18,3 +18,4 @@ void initFunctions_14();
 void initFunctions_15();
 void initFunctions_16();
 void initFunctions_17();
+void initFunctions_18();
