+++ modules/dasClangBind/src/dasClangBind.func.reg.inc
@@ -18,3 +18,4 @@ initFunctions_14();
 initFunctions_15();
 initFunctions_16();
 initFunctions_17();
+initFunctions_18();
