+++ modules/dasClangBind/src/dasClangBind.func_10.cpp
@@ -12,85 +12,85 @@
 namespace das {
 #include "dasClangBind.func.aot.decl.inc"
 void Module_dasClangBind::initFunctions_10() {
-// from clang-c/Index.h:3487:25
+// from clang-c/Index.h:3744:26
+	makeExtern< long long (*)(CXCursor) , clang_Cursor_getOffsetOfField , SimNode_ExtFuncCall >(lib,"clang_Cursor_getOffsetOfField","clang_Cursor_getOffsetOfField")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:3750:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_Cursor_isAnonymous , SimNode_ExtFuncCall >(lib,"clang_Cursor_isAnonymous","clang_Cursor_isAnonymous")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:3756:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_Cursor_isAnonymousRecordDecl , SimNode_ExtFuncCall >(lib,"clang_Cursor_isAnonymousRecordDecl","clang_Cursor_isAnonymousRecordDecl")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:3762:25
 	makeExtern< unsigned int (*)(CXCursor) , clang_Cursor_isInlineNamespace , SimNode_ExtFuncCall >(lib,"clang_Cursor_isInlineNamespace","clang_Cursor_isInlineNamespace")
 		->args({"C"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3502:20
+// from clang-c/Index.h:3777:20
 	makeExtern< int (*)(CXType) , clang_Type_getNumTemplateArguments , SimNode_ExtFuncCall >(lib,"clang_Type_getNumTemplateArguments","clang_Type_getNumTemplateArguments")
 		->args({"T"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3511:23
+// from clang-c/Index.h:3786:23
 	makeExtern< CXType (*)(CXType,unsigned int) , clang_Type_getTemplateArgumentAsType , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_Type_getTemplateArgumentAsType","clang_Type_getTemplateArgumentAsType")
 		->args({"T","i"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3520:40
+// from clang-c/Index.h:3795:40
 	makeExtern< CXRefQualifierKind (*)(CXType) , clang_Type_getCXXRefQualifier , SimNode_ExtFuncCall >(lib,"clang_Type_getCXXRefQualifier","clang_Type_getCXXRefQualifier")
 		->args({"T"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3526:25
-	makeExtern< unsigned int (*)(CXCursor) , clang_Cursor_isBitField , SimNode_ExtFuncCall >(lib,"clang_Cursor_isBitField","clang_Cursor_isBitField")
-		->args({"C"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3532:25
+// from clang-c/Index.h:3801:25
 	makeExtern< unsigned int (*)(CXCursor) , clang_isVirtualBase , SimNode_ExtFuncCall >(lib,"clang_isVirtualBase","clang_isVirtualBase")
 		->args({""})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3552:43
+// from clang-c/Index.h:3810:26
+	makeExtern< long long (*)(CXCursor,CXCursor) , clang_getOffsetOfBase , SimNode_ExtFuncCall >(lib,"clang_getOffsetOfBase","clang_getOffsetOfBase")
+		->args({"Parent","Base"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:3830:43
 	makeExtern< CX_CXXAccessSpecifier (*)(CXCursor) , clang_getCXXAccessSpecifier , SimNode_ExtFuncCall >(lib,"clang_getCXXAccessSpecifier","clang_getCXXAccessSpecifier")
 		->args({""})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3575:37
+// from clang-c/Index.h:3894:1
+	makeExtern< CX_BinaryOperatorKind (*)(CXCursor) , clang_Cursor_getBinaryOpcode , SimNode_ExtFuncCall >(lib,"clang_Cursor_getBinaryOpcode","clang_Cursor_getBinaryOpcode")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:3902:1
+	makeExtern< CXString (*)(CX_BinaryOperatorKind) , clang_Cursor_getBinaryOpcodeStr , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_Cursor_getBinaryOpcodeStr","clang_Cursor_getBinaryOpcodeStr")
+		->args({"Op"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:3910:37
 	makeExtern< CX_StorageClass (*)(CXCursor) , clang_Cursor_getStorageClass , SimNode_ExtFuncCall >(lib,"clang_Cursor_getStorageClass","clang_Cursor_getStorageClass")
 		->args({""})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3586:25
+// from clang-c/Index.h:3921:25
 	makeExtern< unsigned int (*)(CXCursor) , clang_getNumOverloadedDecls , SimNode_ExtFuncCall >(lib,"clang_getNumOverloadedDecls","clang_getNumOverloadedDecls")
 		->args({"cursor"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3602:25
+// from clang-c/Index.h:3937:25
 	makeExtern< CXCursor (*)(CXCursor,unsigned int) , clang_getOverloadedDecl , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getOverloadedDecl","clang_getOverloadedDecl")
 		->args({"cursor","index"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3620:23
+// from clang-c/Index.h:3955:23
 	makeExtern< CXType (*)(CXCursor) , clang_getIBOutletCollectionType , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getIBOutletCollectionType","clang_getIBOutletCollectionType")
 		->args({""})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3748:25
+// from clang-c/Index.h:4057:1
+	makeExtern< unsigned int (*)(CXCursor,_CXChildVisitResult *) , clang_visitChildrenWithBlock , SimNode_ExtFuncCall >(lib,"clang_visitChildrenWithBlock","clang_visitChildrenWithBlock")
+		->args({"parent","block"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4083:25
 	makeExtern< CXString (*)(CXCursor) , clang_getCursorUSR , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCursorUSR","clang_getCursorUSR")
 		->args({""})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3753:25
+// from clang-c/Index.h:4088:25
 	makeExtern< CXString (*)(const char *) , clang_constructUSR_ObjCClass , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_constructUSR_ObjCClass","clang_constructUSR_ObjCClass")
 		->args({"class_name"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3758:25
+// from clang-c/Index.h:4093:25
 	makeExtern< CXString (*)(const char *,const char *) , clang_constructUSR_ObjCCategory , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_constructUSR_ObjCCategory","clang_constructUSR_ObjCCategory")
 		->args({"class_name","category_name"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3765:1
-	makeExtern< CXString (*)(const char *) , clang_constructUSR_ObjCProtocol , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_constructUSR_ObjCProtocol","clang_constructUSR_ObjCProtocol")
-		->args({"protocol_name"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3771:25
-	makeExtern< CXString (*)(const char *,CXString) , clang_constructUSR_ObjCIvar , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_constructUSR_ObjCIvar","clang_constructUSR_ObjCIvar")
-		->args({"name","classUSR"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3778:25
-	makeExtern< CXString (*)(const char *,unsigned int,CXString) , clang_constructUSR_ObjCMethod , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_constructUSR_ObjCMethod","clang_constructUSR_ObjCMethod")
-		->args({"name","isInstanceMethod","classUSR"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3786:25
-	makeExtern< CXString (*)(const char *,CXString) , clang_constructUSR_ObjCProperty , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_constructUSR_ObjCProperty","clang_constructUSR_ObjCProperty")
-		->args({"property","classUSR"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3792:25
-	makeExtern< CXString (*)(CXCursor) , clang_getCursorSpelling , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCursorSpelling","clang_getCursorSpelling")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:3805:30
-	makeExtern< CXSourceRange (*)(CXCursor,unsigned int,unsigned int) , clang_Cursor_getSpellingNameRange , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_Cursor_getSpellingNameRange","clang_Cursor_getSpellingNameRange")
-		->args({"","pieceIndex","options"})
 		->addToModule(*this, SideEffects::worstDefault);
 }
 }
