+++ modules/dasClangBind/src/dasClangBind.func_14.cpp
@@ -12,83 +12,85 @@
 namespace das {
 #include "dasClangBind.func.aot.decl.inc"
 void Module_dasClangBind::initFunctions_14() {
-// from clang-c/Index.h:4605:21
+// from clang-c/Index.h:4765:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_CXXMethod_isStatic , SimNode_ExtFuncCall >(lib,"clang_CXXMethod_isStatic","clang_CXXMethod_isStatic")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4772:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_CXXMethod_isVirtual , SimNode_ExtFuncCall >(lib,"clang_CXXMethod_isVirtual","clang_CXXMethod_isVirtual")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4797:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_CXXMethod_isCopyAssignmentOperator , SimNode_ExtFuncCall >(lib,"clang_CXXMethod_isCopyAssignmentOperator","clang_CXXMethod_isCopyAssignmentOperator")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4822:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_CXXMethod_isMoveAssignmentOperator , SimNode_ExtFuncCall >(lib,"clang_CXXMethod_isMoveAssignmentOperator","clang_CXXMethod_isMoveAssignmentOperator")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4867:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_CXXMethod_isExplicit , SimNode_ExtFuncCall >(lib,"clang_CXXMethod_isExplicit","clang_CXXMethod_isExplicit")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4873:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_CXXRecord_isAbstract , SimNode_ExtFuncCall >(lib,"clang_CXXRecord_isAbstract","clang_CXXRecord_isAbstract")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4878:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_EnumDecl_isScoped , SimNode_ExtFuncCall >(lib,"clang_EnumDecl_isScoped","clang_EnumDecl_isScoped")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4884:25
+	makeExtern< unsigned int (*)(CXCursor) , clang_CXXMethod_isConst , SimNode_ExtFuncCall >(lib,"clang_CXXMethod_isConst","clang_CXXMethod_isConst")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4903:34
+	makeExtern< CXCursorKind (*)(CXCursor) , clang_getTemplateCursorKind , SimNode_ExtFuncCall >(lib,"clang_getTemplateCursorKind","clang_getTemplateCursorKind")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4933:25
+	makeExtern< CXCursor (*)(CXCursor) , clang_getSpecializedCursorTemplate , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getSpecializedCursorTemplate","clang_getSpecializedCursorTemplate")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:4953:30
+	makeExtern< CXSourceRange (*)(CXCursor,unsigned int,unsigned int) , clang_getCursorReferenceNameRange , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCursorReferenceNameRange","clang_getCursorReferenceNameRange")
+		->args({"C","NameFlags","PieceIndex"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5045:25
+	makeExtern< CXToken * (*)(CXTranslationUnitImpl *,CXSourceLocation) , clang_getToken , SimNode_ExtFuncCall >(lib,"clang_getToken","clang_getToken")
+		->args({"TU","Location"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5051:28
+	makeExtern< CXTokenKind (*)(CXToken) , clang_getTokenKind , SimNode_ExtFuncCall >(lib,"clang_getTokenKind","clang_getTokenKind")
+		->args({""})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5059:25
+	makeExtern< CXString (*)(CXTranslationUnitImpl *,CXToken) , clang_getTokenSpelling , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getTokenSpelling","clang_getTokenSpelling")
+		->args({"",""})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5064:33
+	makeExtern< CXSourceLocation (*)(CXTranslationUnitImpl *,CXToken) , clang_getTokenLocation , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getTokenLocation","clang_getTokenLocation")
+		->args({"",""})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5070:30
+	makeExtern< CXSourceRange (*)(CXTranslationUnitImpl *,CXToken) , clang_getTokenExtent , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getTokenExtent","clang_getTokenExtent")
+		->args({"",""})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5089:21
+	makeExtern< void (*)(CXTranslationUnitImpl *,CXSourceRange,CXToken **,unsigned int *) , clang_tokenize , SimNode_ExtFuncCall >(lib,"clang_tokenize","clang_tokenize")
+		->args({"TU","Range","Tokens","NumTokens"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5122:21
+	makeExtern< void (*)(CXTranslationUnitImpl *,CXToken *,unsigned int,CXCursor *) , clang_annotateTokens , SimNode_ExtFuncCall >(lib,"clang_annotateTokens","clang_annotateTokens")
+		->args({"TU","Tokens","NumTokens","Cursors"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5128:21
 	makeExtern< void (*)(CXTranslationUnitImpl *,CXToken *,unsigned int) , clang_disposeTokens , SimNode_ExtFuncCall >(lib,"clang_disposeTokens","clang_disposeTokens")
 		->args({"TU","Tokens","NumTokens"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4622:25
+// from clang-c/Index.h:5145:25
 	makeExtern< CXString (*)(CXCursorKind) , clang_getCursorKindSpelling , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCursorKindSpelling","clang_getCursorKindSpelling")
 		->args({"Kind"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4623:21
-	makeExtern< void (*)(CXCursor,const char **,const char **,unsigned int *,unsigned int *,unsigned int *,unsigned int *) , clang_getDefinitionSpellingAndExtent , SimNode_ExtFuncCall >(lib,"clang_getDefinitionSpellingAndExtent","clang_getDefinitionSpellingAndExtent")
-		->args({"","startBuf","endBuf","startLine","startColumn","endLine","endColumn"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4626:21
-	makeExtern< void (*)() , clang_enableStackTraces , SimNode_ExtFuncCall >(lib,"clang_enableStackTraces","clang_enableStackTraces")
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4865:1
-	makeExtern< CXCompletionChunkKind (*)(void *,unsigned int) , clang_getCompletionChunkKind , SimNode_ExtFuncCall >(lib,"clang_getCompletionChunkKind","clang_getCompletionChunkKind")
-		->args({"completion_string","chunk_number"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4878:25
-	makeExtern< CXString (*)(void *,unsigned int) , clang_getCompletionChunkText , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCompletionChunkText","clang_getCompletionChunkText")
-		->args({"completion_string","chunk_number"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4892:35
-	makeExtern< void * (*)(void *,unsigned int) , clang_getCompletionChunkCompletionString , SimNode_ExtFuncCall >(lib,"clang_getCompletionChunkCompletionString","clang_getCompletionChunkCompletionString")
-		->args({"completion_string","chunk_number"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4899:1
-	makeExtern< unsigned int (*)(void *) , clang_getNumCompletionChunks , SimNode_ExtFuncCall >(lib,"clang_getNumCompletionChunks","clang_getNumCompletionChunks")
-		->args({"completion_string"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4914:1
-	makeExtern< unsigned int (*)(void *) , clang_getCompletionPriority , SimNode_ExtFuncCall >(lib,"clang_getCompletionPriority","clang_getCompletionPriority")
-		->args({"completion_string"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4925:1
-	makeExtern< CXAvailabilityKind (*)(void *) , clang_getCompletionAvailability , SimNode_ExtFuncCall >(lib,"clang_getCompletionAvailability","clang_getCompletionAvailability")
-		->args({"completion_string"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4937:1
-	makeExtern< unsigned int (*)(void *) , clang_getCompletionNumAnnotations , SimNode_ExtFuncCall >(lib,"clang_getCompletionNumAnnotations","clang_getCompletionNumAnnotations")
-		->args({"completion_string"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4950:25
-	makeExtern< CXString (*)(void *,unsigned int) , clang_getCompletionAnnotation , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCompletionAnnotation","clang_getCompletionAnnotation")
-		->args({"completion_string","annotation_number"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4969:25
-	makeExtern< CXString (*)(void *,CXCursorKind *) , clang_getCompletionParent , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCompletionParent","clang_getCompletionParent")
-		->args({"completion_string","kind"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4977:1
-	makeExtern< CXString (*)(void *) , clang_getCompletionBriefComment , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCompletionBriefComment","clang_getCompletionBriefComment")
-		->args({"completion_string"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:4989:1
-	makeExtern< void * (*)(CXCursor) , clang_getCursorCompletionString , SimNode_ExtFuncCall >(lib,"clang_getCursorCompletionString","clang_getCursorCompletionString")
-		->args({"cursor"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5025:1
-	makeExtern< unsigned int (*)(CXCodeCompleteResults *,unsigned int) , clang_getCompletionNumFixIts , SimNode_ExtFuncCall >(lib,"clang_getCompletionNumFixIts","clang_getCompletionNumFixIts")
-		->args({"results","completion_index"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5071:25
-	makeExtern< CXString (*)(CXCodeCompleteResults *,unsigned int,unsigned int,CXSourceRange *) , clang_getCompletionFixIt , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getCompletionFixIt","clang_getCompletionFixIt")
-		->args({"results","completion_index","fixit_index","replacement_range"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5255:25
-	makeExtern< unsigned int (*)() , clang_defaultCodeCompleteOptions , SimNode_ExtFuncCall >(lib,"clang_defaultCodeCompleteOptions","clang_defaultCodeCompleteOptions")
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5327:1
-	makeExtern< CXCodeCompleteResults * (*)(CXTranslationUnitImpl *,const char *,unsigned int,unsigned int,CXUnsavedFile *,unsigned int,unsigned int) , clang_codeCompleteAt , SimNode_ExtFuncCall >(lib,"clang_codeCompleteAt","clang_codeCompleteAt")
-		->args({"TU","complete_filename","complete_line","complete_column","unsaved_files","num_unsaved_files","options"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5340:6
-	makeExtern< void (*)(CXCompletionResult *,unsigned int) , clang_sortCodeCompletionResults , SimNode_ExtFuncCall >(lib,"clang_sortCodeCompletionResults","clang_sortCodeCompletionResults")
-		->args({"Results","NumResults"})
 		->addToModule(*this, SideEffects::worstDefault);
 }
 }
