+++ modules/dasClangBind/src/dasClangBind.func_16.cpp
@@ -12,85 +12,84 @@
 namespace das {
 #include "dasClangBind.func.aot.decl.inc"
 void Module_dasClangBind::initFunctions_16() {
-// from clang-c/Index.h:5594:25
-	makeExtern< unsigned int (*)(void *) , clang_remap_getNumFiles , SimNode_ExtFuncCall >(lib,"clang_remap_getNumFiles","clang_remap_getNumFiles")
-		->args({""})
+// from clang-c/Index.h:5889:14
+	makeExtern< void * (*)(CXCodeCompleteResults *,unsigned int) , clang_codeCompleteGetDiagnostic , SimNode_ExtFuncCall >(lib,"clang_codeCompleteGetDiagnostic","clang_codeCompleteGetDiagnostic")
+		->args({"Results","Index"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5604:21
-	makeExtern< void (*)(void *,unsigned int,CXString *,CXString *) , clang_remap_getFilenames , SimNode_ExtFuncCall >(lib,"clang_remap_getFilenames","clang_remap_getFilenames")
-		->args({"","index","original","transformed"})
+// from clang-c/Index.h:5903:1
+	makeExtern< unsigned long long (*)(CXCodeCompleteResults *) , clang_codeCompleteGetContexts , SimNode_ExtFuncCall >(lib,"clang_codeCompleteGetContexts","clang_codeCompleteGetContexts")
+		->args({"Results"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5611:21
-	makeExtern< void (*)(void *) , clang_remap_dispose , SimNode_ExtFuncCall >(lib,"clang_remap_dispose","clang_remap_dispose")
-		->args({""})
+// from clang-c/Index.h:5923:1
+	makeExtern< CXCursorKind (*)(CXCodeCompleteResults *,unsigned int *) , clang_codeCompleteGetContainerKind , SimNode_ExtFuncCall >(lib,"clang_codeCompleteGetContainerKind","clang_codeCompleteGetContainerKind")
+		->args({"Results","IsIncomplete"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5660:25
+// from clang-c/Index.h:5936:10
+	makeExtern< CXString (*)(CXCodeCompleteResults *) , clang_codeCompleteGetContainerUSR , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_codeCompleteGetContainerUSR","clang_codeCompleteGetContainerUSR")
+		->args({"Results"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5950:10
+	makeExtern< CXString (*)(CXCodeCompleteResults *) , clang_codeCompleteGetObjCSelector , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_codeCompleteGetObjCSelector","clang_codeCompleteGetObjCSelector")
+		->args({"Results"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5966:25
+	makeExtern< CXString (*)() , clang_getClangVersion , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getClangVersion","clang_getClangVersion")
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:5974:21
+	makeExtern< void (*)(unsigned int) , clang_toggleCrashRecovery , SimNode_ExtFuncCall >(lib,"clang_toggleCrashRecovery","clang_toggleCrashRecovery")
+		->args({"isEnabled"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6025:29
+	makeExtern< void * (*)(CXCursor) , clang_Cursor_Evaluate , SimNode_ExtFuncCall >(lib,"clang_Cursor_Evaluate","clang_Cursor_Evaluate")
+		->args({"C"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6030:33
+	makeExtern< CXEvalResultKind (*)(void *) , clang_EvalResult_getKind , SimNode_ExtFuncCall >(lib,"clang_EvalResult_getKind","clang_EvalResult_getKind")
+		->args({"E"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6036:20
+	makeExtern< int (*)(void *) , clang_EvalResult_getAsInt , SimNode_ExtFuncCall >(lib,"clang_EvalResult_getAsInt","clang_EvalResult_getAsInt")
+		->args({"E"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6043:26
+	makeExtern< long long (*)(void *) , clang_EvalResult_getAsLongLong , SimNode_ExtFuncCall >(lib,"clang_EvalResult_getAsLongLong","clang_EvalResult_getAsLongLong")
+		->args({"E"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6049:25
+	makeExtern< unsigned int (*)(void *) , clang_EvalResult_isUnsignedInt , SimNode_ExtFuncCall >(lib,"clang_EvalResult_isUnsignedInt","clang_EvalResult_isUnsignedInt")
+		->args({"E"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6056:1
+	makeExtern< unsigned long long (*)(void *) , clang_EvalResult_getAsUnsigned , SimNode_ExtFuncCall >(lib,"clang_EvalResult_getAsUnsigned","clang_EvalResult_getAsUnsigned")
+		->args({"E"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6062:23
+	makeExtern< double (*)(void *) , clang_EvalResult_getAsDouble , SimNode_ExtFuncCall >(lib,"clang_EvalResult_getAsDouble","clang_EvalResult_getAsDouble")
+		->args({"E"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6070:28
+	makeExtern< const char * (*)(void *) , clang_EvalResult_getAsStr , SimNode_ExtFuncCall >(lib,"clang_EvalResult_getAsStr","clang_EvalResult_getAsStr")
+		->args({"E"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6075:21
+	makeExtern< void (*)(void *) , clang_EvalResult_dispose , SimNode_ExtFuncCall >(lib,"clang_EvalResult_dispose","clang_EvalResult_dispose")
+		->args({"E"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:6123:25
 	makeExtern< CXResult (*)(CXCursor,void *,CXCursorAndRangeVisitor) , clang_findReferencesInFile , SimNode_ExtFuncCall >(lib,"clang_findReferencesInFile","clang_findReferencesInFile")
 		->args({"cursor","file","visitor"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:5675:25
+// from clang-c/Index.h:6138:25
 	makeExtern< CXResult (*)(CXTranslationUnitImpl *,void *,CXCursorAndRangeVisitor) , clang_findIncludesInFile , SimNode_ExtFuncCall >(lib,"clang_findIncludesInFile","clang_findIncludesInFile")
 		->args({"TU","file","visitor"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6077:20
-	makeExtern< int (*)(CXIdxEntityKind) , clang_index_isEntityObjCContainerKind , SimNode_ExtFuncCall >(lib,"clang_index_isEntityObjCContainerKind","clang_index_isEntityObjCContainerKind")
-		->args({""})
+// from clang-c/Index.h:6149:10
+	makeExtern< CXResult (*)(CXCursor,void *,_CXCursorAndRangeVisitorBlock *) , clang_findReferencesInFileWithBlock , SimNode_ExtFuncCall >(lib,"clang_findReferencesInFileWithBlock","clang_findReferencesInFileWithBlock")
+		->args({"","",""})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6079:1
-	makeExtern< const CXIdxObjCContainerDeclInfo * (*)(const CXIdxDeclInfo *) , clang_index_getObjCContainerDeclInfo , SimNode_ExtFuncCall >(lib,"clang_index_getObjCContainerDeclInfo","clang_index_getObjCContainerDeclInfo")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6082:1
-	makeExtern< const CXIdxObjCInterfaceDeclInfo * (*)(const CXIdxDeclInfo *) , clang_index_getObjCInterfaceDeclInfo , SimNode_ExtFuncCall >(lib,"clang_index_getObjCInterfaceDeclInfo","clang_index_getObjCInterfaceDeclInfo")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6086:1
-	makeExtern< const CXIdxObjCCategoryDeclInfo * (*)(const CXIdxDeclInfo *) , clang_index_getObjCCategoryDeclInfo , SimNode_ExtFuncCall >(lib,"clang_index_getObjCCategoryDeclInfo","clang_index_getObjCCategoryDeclInfo")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6089:1
-	makeExtern< const CXIdxObjCProtocolRefListInfo * (*)(const CXIdxDeclInfo *) , clang_index_getObjCProtocolRefListInfo , SimNode_ExtFuncCall >(lib,"clang_index_getObjCProtocolRefListInfo","clang_index_getObjCProtocolRefListInfo")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6092:1
-	makeExtern< const CXIdxObjCPropertyDeclInfo * (*)(const CXIdxDeclInfo *) , clang_index_getObjCPropertyDeclInfo , SimNode_ExtFuncCall >(lib,"clang_index_getObjCPropertyDeclInfo","clang_index_getObjCPropertyDeclInfo")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6095:1
-	makeExtern< const CXIdxIBOutletCollectionAttrInfo * (*)(const CXIdxAttrInfo *) , clang_index_getIBOutletCollectionAttrInfo , SimNode_ExtFuncCall >(lib,"clang_index_getIBOutletCollectionAttrInfo","clang_index_getIBOutletCollectionAttrInfo")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6098:1
-	makeExtern< const CXIdxCXXClassDeclInfo * (*)(const CXIdxDeclInfo *) , clang_index_getCXXClassDeclInfo , SimNode_ExtFuncCall >(lib,"clang_index_getCXXClassDeclInfo","clang_index_getCXXClassDeclInfo")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6105:1
-	makeExtern< void * (*)(const CXIdxContainerInfo *) , clang_index_getClientContainer , SimNode_ExtFuncCall >(lib,"clang_index_getClientContainer","clang_index_getClientContainer")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6111:21
-	makeExtern< void (*)(const CXIdxContainerInfo *,void *) , clang_index_setClientContainer , SimNode_ExtFuncCall >(lib,"clang_index_setClientContainer","clang_index_setClientContainer")
-		->args({"",""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6118:1
-	makeExtern< void * (*)(const CXIdxEntityInfo *) , clang_index_getClientEntity , SimNode_ExtFuncCall >(lib,"clang_index_getClientEntity","clang_index_getClientEntity")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6123:21
-	makeExtern< void (*)(const CXIdxEntityInfo *,void *) , clang_index_setClientEntity , SimNode_ExtFuncCall >(lib,"clang_index_setClientEntity","clang_index_setClientEntity")
-		->args({"",""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6138:30
-	makeExtern< void * (*)(void *) , clang_IndexAction_create , SimNode_ExtFuncCall >(lib,"clang_IndexAction_create","clang_IndexAction_create")
-		->args({"CIdx"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6146:21
-	makeExtern< void (*)(void *) , clang_IndexAction_dispose , SimNode_ExtFuncCall >(lib,"clang_IndexAction_dispose","clang_IndexAction_dispose")
-		->args({""})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:6212:20
-	makeExtern< int (*)(void *,void *,IndexerCallbacks *,unsigned int,unsigned int,const char *,const char *const *,int,CXUnsavedFile *,unsigned int,CXTranslationUnitImpl **,unsigned int) , clang_indexSourceFile , SimNode_ExtFuncCall >(lib,"clang_indexSourceFile","clang_indexSourceFile")
-		->args({"","client_data","index_callbacks","index_callbacks_size","index_options","source_filename","command_line_args","num_command_line_args","unsaved_files","num_unsaved_files","out_TU","TU_options"})
+// from clang-c/Index.h:6153:10
+	makeExtern< CXResult (*)(CXTranslationUnitImpl *,void *,_CXCursorAndRangeVisitorBlock *) , clang_findIncludesInFileWithBlock , SimNode_ExtFuncCall >(lib,"clang_findIncludesInFileWithBlock","clang_findIncludesInFileWithBlock")
+		->args({"","",""})
 		->addToModule(*this, SideEffects::worstDefault);
 }
 }
