+++ modules/dasClangBind/src/dasClangBind.func_3.cpp
@@ -12,6 +12,10 @@
 namespace das {
 #include "dasClangBind.func.aot.decl.inc"
 void Module_dasClangBind::initFunctions_3() {
+// from clang-c/CXDiagnostic.h:321:25
+	makeExtern< unsigned int (*)(void *) , clang_getDiagnosticNumRanges , SimNode_ExtFuncCall >(lib,"clang_getDiagnosticNumRanges","clang_getDiagnosticNumRanges")
+		->args({""})
+		->addToModule(*this, SideEffects::worstDefault);
 // from clang-c/CXDiagnostic.h:336:30
 	makeExtern< CXSourceRange (*)(void *,unsigned int) , clang_getDiagnosticRange , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getDiagnosticRange","clang_getDiagnosticRange")
 		->args({"Diagnostic","Range"})
@@ -24,73 +28,69 @@ void Module_dasClangBind::initFunctions_3() {
 	makeExtern< CXString (*)(void *,unsigned int,CXSourceRange *) , clang_getDiagnosticFixIt , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getDiagnosticFixIt","clang_getDiagnosticFixIt")
 		->args({"Diagnostic","FixIt","ReplacementRange"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:267:24
+// from clang-c/Index.h:271:24
 	makeExtern< void * (*)(int,int) , clang_createIndex , SimNode_ExtFuncCall >(lib,"clang_createIndex","clang_createIndex")
 		->args({"excludeDeclarationsFromPCH","displayDiagnostics"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:276:21
+// from clang-c/Index.h:280:21
 	makeExtern< void (*)(void *) , clang_disposeIndex , SimNode_ExtFuncCall >(lib,"clang_disposeIndex","clang_disposeIndex")
 		->args({"index"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:325:21
+// from clang-c/Index.h:447:1
+	makeExtern< void * (*)(const CXIndexOptions *) , clang_createIndexWithOptions , SimNode_ExtFuncCall >(lib,"clang_createIndexWithOptions","clang_createIndexWithOptions")
+		->args({"options"})
+		->addToModule(*this, SideEffects::worstDefault);
+// from clang-c/Index.h:467:21
 	makeExtern< void (*)(void *,unsigned int) , clang_CXIndex_setGlobalOptions , SimNode_ExtFuncCall >(lib,"clang_CXIndex_setGlobalOptions","clang_CXIndex_setGlobalOptions")
 		->args({"","options"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:333:25
+// from clang-c/Index.h:478:25
 	makeExtern< unsigned int (*)(void *) , clang_CXIndex_getGlobalOptions , SimNode_ExtFuncCall >(lib,"clang_CXIndex_getGlobalOptions","clang_CXIndex_getGlobalOptions")
 		->args({""})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:343:1
+// from clang-c/Index.h:491:1
 	makeExtern< void (*)(void *,const char *) , clang_CXIndex_setInvocationEmissionPathOption , SimNode_ExtFuncCall >(lib,"clang_CXIndex_setInvocationEmissionPathOption","clang_CXIndex_setInvocationEmissionPathOption")
 		->args({"","Path"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:350:25
+// from clang-c/Index.h:498:25
 	makeExtern< unsigned int (*)(CXTranslationUnitImpl *,void *) , clang_isFileMultipleIncludeGuarded , SimNode_ExtFuncCall >(lib,"clang_isFileMultipleIncludeGuarded","clang_isFileMultipleIncludeGuarded")
 		->args({"tu","file"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:363:23
+// from clang-c/Index.h:511:23
 	makeExtern< void * (*)(CXTranslationUnitImpl *,const char *) , clang_getFile , SimNode_ExtFuncCall >(lib,"clang_getFile","clang_getFile")
 		->args({"tu","file_name"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:378:28
+// from clang-c/Index.h:526:28
 	makeExtern< const char * (*)(CXTranslationUnitImpl *,void *,size_t *) , clang_getFileContents , SimNode_ExtFuncCall >(lib,"clang_getFileContents","clang_getFileContents")
 		->args({"tu","file","size"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:385:33
+// from clang-c/Index.h:533:33
 	makeExtern< CXSourceLocation (*)(CXTranslationUnitImpl *,void *,unsigned int,unsigned int) , clang_getLocation , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getLocation","clang_getLocation")
 		->args({"tu","file","line","column"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:392:33
+// from clang-c/Index.h:540:33
 	makeExtern< CXSourceLocation (*)(CXTranslationUnitImpl *,void *,unsigned int) , clang_getLocationForOffset , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getLocationForOffset","clang_getLocationForOffset")
 		->args({"tu","file","offset"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:402:35
+// from clang-c/Index.h:550:35
 	makeExtern< CXSourceRangeList * (*)(CXTranslationUnitImpl *,void *) , clang_getSkippedRanges , SimNode_ExtFuncCall >(lib,"clang_getSkippedRanges","clang_getSkippedRanges")
 		->args({"tu","file"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:413:1
+// from clang-c/Index.h:561:1
 	makeExtern< CXSourceRangeList * (*)(CXTranslationUnitImpl *) , clang_getAllSkippedRanges , SimNode_ExtFuncCall >(lib,"clang_getAllSkippedRanges","clang_getAllSkippedRanges")
 		->args({"tu"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:419:25
+// from clang-c/Index.h:567:25
 	makeExtern< unsigned int (*)(CXTranslationUnitImpl *) , clang_getNumDiagnostics , SimNode_ExtFuncCall >(lib,"clang_getNumDiagnostics","clang_getNumDiagnostics")
 		->args({"Unit"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:430:29
+// from clang-c/Index.h:578:29
 	makeExtern< void * (*)(CXTranslationUnitImpl *,unsigned int) , clang_getDiagnostic , SimNode_ExtFuncCall >(lib,"clang_getDiagnostic","clang_getDiagnostic")
 		->args({"Unit","Index"})
 		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:440:1
+// from clang-c/Index.h:588:1
 	makeExtern< void * (*)(CXTranslationUnitImpl *) , clang_getDiagnosticSetFromTU , SimNode_ExtFuncCall >(lib,"clang_getDiagnosticSetFromTU","clang_getDiagnosticSetFromTU")
 		->args({"Unit"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:456:1
-	makeExtern< CXString (*)(CXTranslationUnitImpl *) , clang_getTranslationUnitSpelling , SimNode_ExtFuncCallAndCopyOrMove >(lib,"clang_getTranslationUnitSpelling","clang_getTranslationUnitSpelling")
-		->args({"CTUnit"})
-		->addToModule(*this, SideEffects::worstDefault);
-// from clang-c/Index.h:498:34
-	makeExtern< CXTranslationUnitImpl * (*)(void *,const char *,int,const char *const *,unsigned int,CXUnsavedFile *) , clang_createTranslationUnitFromSourceFile , SimNode_ExtFuncCall >(lib,"clang_createTranslationUnitFromSourceFile","clang_createTranslationUnitFromSourceFile")
-		->args({"CIdx","source_filename","num_clang_command_line_args","clang_command_line_args","num_unsaved_files","unsaved_files"})
 		->addToModule(*this, SideEffects::worstDefault);
 }
 }
