+++ modules/dasClangBind/src/dasClangBind.struct.decl.inc
@@ -11,16 +11,19 @@ MAKE_EXTERNAL_TYPE_FACTORY(CXTargetInfoImpl,CXTargetInfoImpl);
 MAKE_EXTERNAL_TYPE_FACTORY(CXTranslationUnitImpl,CXTranslationUnitImpl);
 MAKE_EXTERNAL_TYPE_FACTORY(CXUnsavedFile,CXUnsavedFile);
 MAKE_EXTERNAL_TYPE_FACTORY(CXVersion,CXVersion);
+MAKE_EXTERNAL_TYPE_FACTORY(CXIndexOptions,CXIndexOptions);
 MAKE_EXTERNAL_TYPE_FACTORY(CXTUResourceUsageEntry,CXTUResourceUsageEntry);
 MAKE_EXTERNAL_TYPE_FACTORY(CXTUResourceUsage,CXTUResourceUsage);
 MAKE_EXTERNAL_TYPE_FACTORY(CXCursor,CXCursor);
 MAKE_EXTERNAL_TYPE_FACTORY(CXPlatformAvailability,CXPlatformAvailability);
 MAKE_EXTERNAL_TYPE_FACTORY(CXCursorSetImpl,CXCursorSetImpl);
 MAKE_EXTERNAL_TYPE_FACTORY(CXType,CXType);
+MAKE_EXTERNAL_TYPE_FACTORY(_CXChildVisitResult,_CXChildVisitResult);
 MAKE_EXTERNAL_TYPE_FACTORY(CXToken,CXToken);
 MAKE_EXTERNAL_TYPE_FACTORY(CXCompletionResult,CXCompletionResult);
 MAKE_EXTERNAL_TYPE_FACTORY(CXCodeCompleteResults,CXCodeCompleteResults);
 MAKE_EXTERNAL_TYPE_FACTORY(CXCursorAndRangeVisitor,CXCursorAndRangeVisitor);
+MAKE_EXTERNAL_TYPE_FACTORY(_CXCursorAndRangeVisitorBlock,_CXCursorAndRangeVisitorBlock);
 MAKE_EXTERNAL_TYPE_FACTORY(CXIdxLoc,CXIdxLoc);
 MAKE_EXTERNAL_TYPE_FACTORY(CXIdxIncludedFileInfo,CXIdxIncludedFileInfo);
 MAKE_EXTERNAL_TYPE_FACTORY(CXIdxImportedASTFileInfo,CXIdxImportedASTFileInfo);
