+++ modules/dasClangBind/src/dasClangBind.struct.impl.inc
@@ -11,16 +11,19 @@ IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXTargetInfoImpl,CXTargetInfoImpl);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXTranslationUnitImpl,CXTranslationUnitImpl);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXUnsavedFile,CXUnsavedFile);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXVersion,CXVersion);
+IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXIndexOptions,CXIndexOptions);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXTUResourceUsageEntry,CXTUResourceUsageEntry);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXTUResourceUsage,CXTUResourceUsage);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXCursor,CXCursor);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXPlatformAvailability,CXPlatformAvailability);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXCursorSetImpl,CXCursorSetImpl);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXType,CXType);
+IMPLEMENT_EXTERNAL_TYPE_FACTORY(_CXChildVisitResult,_CXChildVisitResult);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXToken,CXToken);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXCompletionResult,CXCompletionResult);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXCodeCompleteResults,CXCodeCompleteResults);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXCursorAndRangeVisitor,CXCursorAndRangeVisitor);
+IMPLEMENT_EXTERNAL_TYPE_FACTORY(_CXCursorAndRangeVisitorBlock,_CXCursorAndRangeVisitorBlock);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXIdxLoc,CXIdxLoc);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXIdxIncludedFileInfo,CXIdxIncludedFileInfo);
 IMPLEMENT_EXTERNAL_TYPE_FACTORY(CXIdxImportedASTFileInfo,CXIdxImportedASTFileInfo);
